requirements.txt
streamlit
rembg
pillow
numpy
